requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dataclasses==0.8
discord.py==1.5.1
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
mypy-extensions==0.4.3
pathlib==1.0.1
pathspec==0.8.1
python-decouple==3.3
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.0
soupsieve==2.0.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.5.1